foundation: moved balance and blockchain objects to token.go
