+++ foundation/token.go
@@ -28,6 +28,16 @@ type Token struct {
 	Blockchain BlockchainName `json:"blockchain"`
 }
 
+type Balance struct {
+	Address string `json:"address"`
+	Amount  string `json:"balance"`
+}
+
+type Blockchain struct {
+	Name    BlockchainName `json:"name"`
+	Address string         `json:"address"`
+}
+
 func (token Token) String() string {
 	return string(token.Name)
 }
